Move interposition into its own module.
