+++ higher-order-effects.cabal
@@ -26,6 +26,7 @@ library
                      , Control.Effect.Fail.Internal
                      , Control.Effect.Fresh
                      , Control.Effect.Internal
+                     , Control.Effect.Interpose
                      , Control.Effect.Lift
                      , Control.Effect.Lift.Internal
                      , Control.Effect.NonDet
