+++ src/Control/Effect.hs
@@ -3,11 +3,12 @@ module Control.Effect
 ) where
 
 import Control.Effect.Carrier   as X (Carrier, Effect)
-import Control.Effect.Eavesdrop as X (Eavesdrop, EavesdropC, Interpose, InterposeC)
+import Control.Effect.Eavesdrop as X (Eavesdrop, EavesdropC)
 import Control.Effect.Error     as X (Error, ErrorC)
 import Control.Effect.Fail      as X (Fail, FailC)
 import Control.Effect.Fresh     as X (Fresh, FreshC)
 import Control.Effect.Internal  as X (Eff, interpret)
+import Control.Effect.Interpose as X (Interpose, InterposeC)
 import Control.Effect.Lift      as X (Lift, LiftC, runM)
 import Control.Effect.NonDet    as X (NonDet, AltC)
 import Control.Effect.Reader    as X (Reader, ReaderC)
